requirements.txt
pytest
selenium
pytest-selenium
python-dotenv
allure-pytest